optimize database tables

+++ app/Database/Migrations/2023-05-17-064955_AsuransiCabang.php
@@ -44,7 +44,7 @@ class AsuransiCabang extends Migration
                 'default' => null
             ),
             'actives' => array(
-                'type' => 'INT',
+                'type' => 'TINYINT',
                 'constraint' => 1,
                 'unsigned' => true,
                 'default' => 0
